--- models/renalStructures_UNEST_segmentation/scripts/networks/nest/layers/__init__.py
@@ -1,11 +1,0 @@
-#!/usr/bin/env python3
-from .conv3d_same import Conv3dSame, conv3d_same
-from .create_conv3d import create_conv3d
-from .drop import DropPath
-from .helpers import to_ntuple
-from .linear import Linear
-from .mlp import Mlp
-from .padding import get_padding, get_same_padding, pad_same
-from .pool3d_same import create_pool3d
-from .trace_utils import _assert
-from .weight_init import trunc_normal_
