--- models/renalStructures_UNEST_segmentation/scripts/networks/nest/layers/conv3d_same.py
@@ -1,46 +1,0 @@
-#!/usr/bin/env python3
-
-""" Conv2d w/ Same Padding
-
-Hacked together by / Copyright 2020 Ross Wightman
-"""
-from typing import Optional, Tuple
-
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-
-from .padding import get_padding_value, pad_same
-
-
-def conv3d_same(
-    x,
-    weight: torch.Tensor,
-    bias: Optional[torch.Tensor] = None,
-    stride: Tuple[int, int] = (1, 1, 1),
-    padding: Tuple[int, int] = (0, 0, 0),
-    dilation: Tuple[int, int] = (1, 1, 1),
-    groups: int = 1,
-):
-    x = pad_same(x, weight.shape[-3:], stride, dilation)
-    return F.conv3d(x, weight, bias, stride, (0, 0, 0), dilation, groups)
-
-
-class Conv3dSame(nn.Conv2d):
-    """Tensorflow like 'SAME' convolution wrapper for 2D convolutions"""
-
-    def __init__(self, in_channels, out_channels, kernel_size, stride=1, padding=0, dilation=1, groups=1, bias=True):
-        super(Conv3dSame, self).__init__(in_channels, out_channels, kernel_size, stride, 0, dilation, groups, bias)
-
-    def forward(self, x):
-        return conv3d_same(x, self.weight, self.bias, self.stride, self.padding, self.dilation, self.groups)
-
-
-def create_conv3d_pad(in_chs, out_chs, kernel_size, **kwargs):
-    padding = kwargs.pop("padding", "")
-    kwargs.setdefault("bias", False)
-    padding, is_dynamic = get_padding_value(padding, kernel_size, **kwargs)
-    if is_dynamic:
-        return Conv3dSame(in_chs, out_chs, kernel_size, **kwargs)
-    else:
-        return nn.Conv3d(in_chs, out_chs, kernel_size, padding=padding, **kwargs)
